% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-get-user-dir.R
\name{spark_get_user_dir}
\alias{spark_get_user_dir}
\title{Get User Directory}
\usage{
spark_get_user_dir(sc)
}
\arguments{
\item{sc}{Spark Context}
}
\value{
User directory path
}
\description{
Gets user directory
}
\examples{
\dontrun{
spark_get_user_dir(sc)
}
}
