% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklyrExtras-package.R
\docType{package}
\name{sparklyrExtras-package}
\alias{sparklyrExtras}
\alias{sparklyrExtras-package}
\title{sparklyrExtras: sparklyrExtras provides a set of simple, unsophisticated convenience 
  functions for working with sparklyr}
\description{
The package provides a fairly liberal collection of potentially
  useful functions for working with Spark.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/konradzdeb/sparklyrExtras/issues}
}

}
\author{
\strong{Maintainer}: Zdeb Konrad \email{konrad.zdeb@me.com}

}
\keyword{internal}
